Change the profiling data structure to log events rather than build graphs.
The online data structure is now simply a list of "events". That list
can be post-processed into a graph structure (equivalent to what we used
to construct online).

The profiling instrumentation stores the list of events in a mutable
box, and updates it with box-cas!, so that streaming infrastructure can
atomically remove events from the list.

This data structure is more efficient (~30% faster on WebSynth) and,
more importantly, much easier to stream.

TODO: streaming is disabled right now; need to adopt this new format.